bugs fixed

+++ My_plugin/ads_each_post.php
@@ -34,14 +34,8 @@
 						break;
 					}
 				}
-
 				if($ad_name_exist){
-					print($ad_name_exist);
 					echo "<script type='text/javascript'> window.alert('Old ad has been overwritten');
-						</script>";
-				}
-				else{
-					echo "<script type='text/javascript'> window.alert('Add new ad);
 						</script>";
 				}
 
@@ -57,7 +51,7 @@
 		echo '<div class="wrap">
 		<h1>New ad befor post</h1>
 		<form name="naph_form" method="post">
-		<input type="hidden" name="naph_do_change" value="Y">Zaawansowane Technologie Webowe - laboratorium 12
+		<input type="hidden" name="naph_do_change" value="Y">Zaawansowane Technologie Webowe - laboratorium 1
 		 	<p>Name of ad: <input id="ad_name" name="ad_name" required></p>
 			<textarea id="html_code" name="html_code" rows="6" cols="50" placeholder="Put your html code of ad here" required></textarea>
 			<p class="submit"><input type="submit" value="Add new advertising"></p>
@@ -73,37 +67,38 @@
 
 		echo '<table class="wp-list-table widefat plugins">
 					<thead>
-					<tr>
-						<td id="cb" class="manage-column column-cb check-column">
-							
-						</td>
-						<th scope="col" id="name" class="manage-column column-name column-primary">Advertising</th>
-						
-						
+					<tr style="background-color: #e5faf7;">
+						<td></td>
+						<th scope="col" id="name" >Advertising</th>
 					</tr>
 					</thead>
-
 					<tbody id="the-list">';
-		foreach ($adds_array as $key => $value) {
-			echo'<tr class="inactive update" data-slug="akismet" data-plugin="akismet/akismet.php">
-						<th scope="row" class="check-column">
-						</th>
-
-						<td class="plugin-title column-primary"><strong>';
-			echo $key;
-			echo	'</strong>
-					<div class="row-actions visible"><span class="activate">
-						
-						<a href="delete_ad.php?delete_ad='.$key.'" id="delete" class="delete" >Usuń</a></span></div>
-
-					</td>
+		if(empty($adds_array)){
+			echo'<tr style="background-color: #f6e7e4;">
+					<th></th>
+					<td class="plugin-title column-primary"><span style="color:red;">There are no ads.</span></td>
 				</tr>';
+		}
+		else{
+			foreach ($adds_array as $key => $value) {
+				echo'<tr style="background-color: #b9fdbf;">
+							<th></th>
+							<td class="plugin-title column-primary"><strong>';
+				echo $key;
+				echo	'</strong>
+						<div class="row-actions visible"><span class="activate">
+							
+							<a href="delete_ad.php?delete_ad='.$key.'" id="delete" class="delete" >Usuń</a></span></div>
 
+						</td>
+					</tr>';
+
+			}
 		}
 		echo '<tfoot>
-					<tr>
-					<td class="manage-column column-cb check-column"></td>
-					<th scope="col" class="manage-column column-name column-primary">Created by Hleb Liaonik, Aliaksandr Nadzeika</th>
+					<tr style="background-color: #e5faf7;">
+						<td></td>
+						<th scope="col" class="manage-column column-name column-primary">Created by Hleb Liaonik, Aliaksandr Nadzeika</th>
 					</tr>
 				</tfoot>
 
@@ -116,18 +111,24 @@
 	add_filter("the_content", "add_content_before",1); 
 	function add_content_before($content){
 		$adds_array = get_option('html_code'); 
-		$random_ad = array_rand($adds_array,1);
 
-		$custom_content = '<div class="entry-content">
-							    <p>';
-		$custom_content .= $adds_array[$random_ad];
-		$custom_content .= '</p></div>';
+		if (!empty($adds_array)) {
+			
+			
+			$random_ad = array_rand($adds_array,1);
 
-							 
-		$custom_content .= $content;
+			$custom_content = '<div class="entry-content">
+								    <p>';
+			$custom_content .= $adds_array[$random_ad];
+			$custom_content .= '</p></div>';
+
+								 
+			$custom_content .= $content;
 
 
-		return $custom_content; 
+			return $custom_content; 
+		}
+		return $content;
 	}
